Split Disc into Sort and Group.
In Fritz's original papers both disc and sdisc sort sums sequentially, but this violates the stable renaming policy where groupings are concerned and importantly it gets in the way of giving the wrong answers for things like `nub :: [Bool] -> [Bool]`.

Now 'Group' uses a different Divisible instance that is careful to properly interleave the equivalence classes. This will enable us, later to make a more productive form of `grouping` for online streaming.